bit machine: re-arrange module
Now the bit machine lives directly in the `bit_machine` module rather
than having an extra `exec::` layer that didn't really serve any
purpose.
